LPS-77392 Rename
